asm generation

+++ type.go
@@ -3,14 +3,14 @@ package main
 import "golang.org/x/exp/slices"
 
 type Type interface {
-	Size() uint
+	Size() uint64
 	Cast(as Type) bool
 }
 
 // Language type primitive, cannot be de-constructed into simpler types.
 // Signedness and size are used to choose the correct CPU Instruction
 type Atom struct {
-	size   uint
+	size   uint64
 	signed bool
 	float  bool
 }
@@ -21,7 +21,7 @@ type Struct struct {
 	Members []Var
 }
 
-func (at Atom) Size() uint {
+func (at Atom) Size() uint64 {
 	return at.size
 }
 
@@ -30,8 +30,8 @@ func (at Atom) Cast(as Type) bool {
 	return same
 }
 
-func (s Struct) Size() uint {
-	size := uint(0)
+func (s Struct) Size() uint64 {
+	size := uint64(0)
 	for _, member := range s.Members {
 		size += member.Type.Size()
 	}
@@ -48,7 +48,7 @@ func (s Struct) Cast(as Type) bool {
 	return false
 }
 
-func (v Void) Size() uint {
+func (v Void) Size() uint64 {
 	return 0
 }
